refactor: moved Course Creator validations to utils
